Check API for details confirmed
 ## Summary
Before we added the green 'save and confirm' button for supplier
details, we had to detect whether or not a supplier's details were
complete from the presence of all the required keys. Now that we have
the button, and an associated flag on the supplier table in the API that
tracks whether the details are confirmed, we should use this instead to
determine whether or not the frontend should display 'your details are
not complete'-esque messages or not.