Add event tracking to analytics js
This has been ported over the buyer app so we can track service copying
when G-10 opens. Ideally this code should be in a shared place, the
frontend-toolkit. This is on Dilwoars list of things to do.

To make it more complicated we intentionally don't share apps
`_events.js` files. See here:
https://github.com/alphagov/digitalmarketplace-frontend-toolkit/commit/1d88cb8c6454c23fe89f96d901320f5935948504