Use framework dates from API
 ## Summary
We currently store dates for important framework lifecycle events as
date strings in a `dates.yml` file stored in
digitalmarketplace-frameworks for each framework. Given it's a string,
it's not ideal for machine processing and makes it hard to format in
different ways. We also have some denormalisation as we store one of the
datetimes in two places: the API and the dates file.

We are now migrating all dates onto the `framework` record on the API
and storing them as real datetimes (returned as ISO8601 datetimestamps
from the API). With this standardised format and centralised source of
truth, we are less likely to have these gets get out of sync and it's
easier to use our existing datetime formatters to present it
appropriately depending on the context.

We normally present each date in a single way across the Digital
Marketplace, so a utility function in dm-utils injects these standard
formats for each date into the `framework` dictionary. These values are
then used during templating rather than the raw UTC values.

This PR also contains a fair bit of work to utilise API stub responses
that are now available in dm-utils to reduce boilerplate and give more
valid test data rather than the smallest subset of keys required for
tests to pass. This should make it easier to write tests in the future
and provide more certainty that the code is functioning correctly.