Fix tests
 ## Summary
Some of the tests for deactivating supplier users incorrectly mocked out
`find_users`, which is not called as part of the view.
`test_cannot_deactivate_nonexistent_id` was missing a mock for
`get_user`, which meant it wasn't really testing the view correctly.
This mock has been added and an assertion included to make sure that the
mock has been called correctly.