Consolidate validation mastheads  ## Summary Use validation masthead from the frontend-toolkit consistently.
Also converts most pages over to using the base page from the frontend-toolkit.

The framework clarification question and signature upload pages have not been completely converted as they contains a rather peculiar form and the views are split in a way that doesn't make them trivial to consolidate (in the time I'm willing to spend on this).

 ## Ticket
https://trello.com/c/RSYEM3FL/448