Add error page for 400s
Unlike the buyer or user frontends, the supplier frontend didn't have a
custom error page, or even any error handling at all, for 400s. This
meant that any time `abort(400)` was being called the user would get
a janky horrible error page.

This allows a custom error message to also be passed through which wil
be displayed to the user.