Change link text for company details
 ## Summary
If a supplier hasn't confirmed their details yet, they can change
everything except their DUNS number themselves, so the link text should
indicate this. Once they have confirmed the details, some of these are
locked in and can only be changed by contacting support. Where this is
the case, we change the link text to 'Correct a mistake' to indicate a
different process and the condition under which a change will be
allowed.