Remove "currently_applying_to" on dashboard visit
 ## Summary
We store a reference to the framework a user is "currently applying to"
when they go to that framework's application dashboard. This is used in
order to show a "return to application" link on the company details
view/confirm page of an application.

If the user visits the company details page from their supplier account
(i.e. not an application), it doesn't make sense to show the link to an
application.