Protect application editing routes with decorator
When applications for a framework close some views are no longer
available. This decorator is applied to those routes to provide users
with meaningful feedback and some extra logging for us.

The apiclient that's passed to the outer wrapper of the decorator is a
callable. This is because the decorator is evaluated at module import
time, which makes mocking the apiclient difficult. By providing it as a
callable it's evaluated at runtime and a mock can be patched in.

The custom error template includes content that is in the frameworks
repo. Its content that describes the expected following framework. If
that hasn't been defined, we show a reduced error message.