Fix cookie warnings
Flask was complaining that our cookie domain was invalid, which resulted
in a lot of warning messages in our pytest output.

It turns out this was because we were setting the `SERVER_NAME` config
variable in the Test config, which as far as I can tell is unnecessary;
none of our other apps do it, and removing the setting didn't break any
tests.