Show expired DOS2 on supplier dashboard
As DOS2 has expired it's no longer shown on the supplier dashboard. This
is a problem, as there are still active DOS2 opportunities that DOS2
suppliers can apply to - and will be for a couple of weeks.

Suppliers need to be able to access the details of the opportunities
they've applied to.

This is a temporary fix and will need proper design for a future proof
solution.