Remove services and documents links from DOS2
The supplier dashboard page currently shows DSO2 even though it's
expired. This is a temporary fix. The reason it was brought back was so
suppliers could see the opportunities they'd applied to. The links for
viewing their services and viewing documents however, return 404's as
the framework is expired.

Making those links work is probably the wrong thing to do, and we've
only seen a user need for suppliers to access their opportunities. This
PR removes the broken links.