Removes upload_framework_agreement route
This commit removes the upload_framework_agreement route. It is a route used for uploading a framework agreement pre-g-cloud 8. As we're about to move to G-Cloud 11, this route is no longer necessary. In the interests of keeping the codebase clean I've removed the route and its corresponding tests. I've also removed the email template, as we're not emailing anyone from this route any more.