Replace mandrill with notify in framework_dashboard view
This commit removes the use of mandrill from the framework_dashboard route. It's been replaced with notify and a for-loop. I've also had to create a new template in the Notify UI for new DOS applications. The ids for these two templates are in a dictionary that's passed to the DMNotifyClient when it's initialised. The templates that mandrill used have been deleted. The tests have been updated to reflect the fact that notify is called.

I've also removed an unnecessary patch in test_framework::test_interest_registered_in_framework_on_post