Remove mandrill from framework variation agreement
This commit switches mandrill for notify when emailing users a confirmation of their agreement to a variation in the framework. This change will require that a developer adds an appropriate template to Notify, or that there is a change in the future to make the template sufficiently generic that we can hard-code it.