Alter tests to use new Notify client
This commit changes existing tests for the `view_contract_variation` view so that they work with the new Notify client. The method we’re testing is instantiated at the top of this class as notify_send_email, so we can use self.notify_send_email to check if it’s called rather than patching all over the place.