Add display session timeout message during applications
We want our users to be aware when filling out long forms that they may
be logged out if they do not interact with the website for longer than
an hour.

This commit adds a simple text message at the top of framework
application forms.

See the [ticket] for more details.

[ticket]: https://trello.com/c/92nGYty4/