Fix details polyfill for service categories question
When we included the GOV.UK Design System we inadvertantly included a
details polyfill that conflicts with our own, causing details elements
to stop working on Internet Explorer.

Although we could use the details polyfill from the design system, it
would involve changing the styling of our checkbox tree element, which
would be inadvisable during the framework application period.

Instead we just remove the javascript for the design system; it turns
out we don't need any of it currently.

In future we might need to include javascript for the design system
module by module rather than using `all.js`.