content loader: implement this as a thread/context-local
the content loader isn't threadsafe, so the easiest avenue towards thread-
enabling an app is probably to supply each thread/context with its own
copy of the content loader. the thread that manages to perform the init_app
also gets to pre-load its content, in the process performing the init-time
content checking that we're used to. other threads will construct their
content loader on first access (it's probably quite tricky to get them to
pre-load too)