content loader: implement this as a thread/context-local
the content loader isn't threadsafe, so the easiest avenue towards thread-
enabling an app is probably to supply each thread/context with its own
copy of the content loader. rather than constructing a new ContentLoader
from scratch for each thread, allow the initial importing process to
construct a "master" ContentLoader which we seal in a closure. the only
things that can escape this closure are deep copies which should not share
any mutable objects with each other and therefore be sufficiently
independent.

each thread will make its own deep copy of this master content loader when
it first needs one - a comparatively cheap operation