Remove DOS1 temporary message
We have a temporary message card that appears only if DOS1 (then just
DOS) is coming.

Since I don't think that will ever happen again, I've replaced it with a
placedholder comment, so we can keep the file, but not have the
temporary message HTML and styles cluttering up the place.