Improve framework dashboard load time for local envs
In order to show the 'last updated' dates on the framework legal documents, we traverse the entire framework directory in the S3 comms bucket. For local developers, this means *everything* in the dev-uploads bucket, which could be hundreds of documents (from functional tests). This change won't affect Preview/Staging/Production as submission documents go to a separate bucket.

Local testing improved page load times from ~30s to ~2s.