Add an override for pending services table on supplier dashboard
We want the border for summary items which are incomplete to cover the
whole row, not just the text span of the first field.

This commit adds an override specifically for the pending services
table, as this is the only place we know of that uses this style.

Thanks @domoscargin for the styles.

Co-authored-by: Brett Kyle <brett.kyle@digital.cabinet-office.gov.uk>