Distinguish G12 recovery suppliers
We had an outage during G12 closing. Some suppliers will be involved in a recovery process to account for the outage causing them to fail to submit their services in time.

We have previously provided a list of the affected supplier IDs via the environment - see https://github.com/alphagov/digitalmarketplace-credentials/pull/359 and https://github.com/alphagov/digitalmarketplace-aws/pull/797. This PR proves that this all works by displaying a placeholder to the affected suppliers.

Currently, the only supplier so configured is our test supplier, so this will not be shown to any real users and is therefore safe for production. The two tests check this.