Create text for time until close of G12 recovery
We show a banner to G12 recovery users (https://github.com/alphagov/digitalmarketplace-supplier-frontend/pull/1297). We now have the final content for this banner. This includes a time until the end in days/hours/minutes/seconds.

Create a (tested) helper function to generate this text.

Hard-code the deadline. This is OK because the code will be removed after the G12 recovery and will not be re-used. We'll need to update with the deadline when we know it.