Set the G12 recovery deadline in the past
We don't yet know when the deadline is, so set it to a dummy date in the far future. However, this means that the number of days remaining decrements every day and breaks the visual regression tests.

Switch to a dummy date in the far past. The banner will now always read '0 seconds remaining'.