Remove 'Add a service' link
This link was added for the G12 recovery when we thought that suppliers were going to be creating services. Now, we know that they're only going to be able to edit/submit existing drafts via the 'Your G-Cloud 12' page. So we want to remove this link, which is no longer needed.