Stop showing broken link to sign agreement
Previously, we were showing this link to G12 recovery suppliers who were not on the framework. If these suppliers tried to follow the link, they would get an error page.

We still want to pretend to be on the framework for the purposes of getting the right things displayed elsewhere. However, by doing the hack just a little later, we can avoid messing up the determination of whether the supplier needs to sign the framework agreement.

Extend the test to test this.