Change submissions and service submissions pages to redirect to services
For G12 recovery suppliers we need to fix breadcrumbs that link to the
"Services" and lot page, and also change where the user ends up after
marking a service as complete.

This commit does this in a fairly brute-force way, by making those urls
always redirect to the `all_services` route; however I think this should
be fine because they will not need to see the pages we are redirecting
away from anyway.