Remove 'signature-upload' endpoint
All frameworks from DOS 5 and G12 onwards will use e-signature.

'signature-endpoint' was part of the manual signature flow used by all frameworks up to DOS 4 and G11. All those frameworks are now closed. So this flow will never be used again. So we can delete this endpoint to reduce the amount of code we need to support.