Load list of G12 recovery draft IDs
We need this so that we can show the G12 recovery suppliers only those drafts which they are able to submit as part of the recovery process.

Load the list of draft IDs. We will use this later to filter the list of drafts that the recovery suppliers can see.

Co-authored-by: Paula Valenca <paula.valenca@digital.cabinet-office.gov.uk>