Show only selected drafts for the G12 recovery
As part of the G12 recovery process, users are only able to submit certain drafts. To avoid confusion and unnecessary work, we're going to hide all other drafts from the G12 recovery process for suppliers.

Only show drafts that are in the allow-list - both submitted and un-submitted.

Tests prove that this is working.