Parse list of IDs from config only once
We have some lists of IDs that we use as allow-lists. These are needed every time certain pages are loaded.

Previously, we did the string parsing every time. Now, perform the string parsing during app initialisation and store the result. This should remove the performance impact of these new allowlists.