Change G12 banner when all drafts complete
In our user research we dicovered that users expected the banner to
change when they had completed all their drafts.

Change the banner to a success message if the user has no
unsubmitted G12 recovery drafts (see #1325).

Content on this banner is not final.

There is G12 recovery logic to show a banner on the
`/frameworks/<string:framework_slug>/services` route, however
I've not updated this as we've now added a new route
(`/all-services`, see #1301) for the G12 suppliers to see
their services and therefore the banner will never be shown
on the `/services` route.

https://trello.com/c/p3BVSpj6/740-2-change-banner-when-all-drafts-complete