Speed up counting unsubmitted G12 draft services
`get_g12_recovery_draft_ids()` takes about 0.1ms, and the previous
version of this code called it for each of a supplier's draft
services. This would be ok for suppliers with only a few drafts,
but would quickly add up for suppliers with over 100.

Additionally, it seems to be faster to convert a list to a set,
and then take the intersection of two sets than it is to iterate
through the list and check membership for each element (see
discussion on #1329).