Remove banner from 'Your G-Cloud-12 services' page
In our user research, we found that the banner caused problems for some users on this page. The banner confused users because it linked back to the same page, and caused them to miss the green flash messages.

Remove the banner text and put it into paragraphs above the appropriate tables.

I've set it up so that the text only appears if the table has some entries, which seems sensible.