Update G12 recovery banner with final content
In #1329 we added logic to the banner so it showed a different
message if the supplier had submitteed all their draft services.
We hadn't decided what the content should be at that time.

In content review, we decided that this behaviour wasn't quite what
we wanted.

Update the banner with the final content, and add more logic so
the banner tells the supplier how many services they've marked as
complete as they go. If the supplier completes all their services,
show a third message.

Note the nested `{% with %}` in the template. Putting
`"service" if g12_recovery.submitted_drafts == 1 else " services"`
in the string concantenation for `message` caused Jinja to only
render the string up to and including service(s). I don't like this
solution very much, but it's the only one I was able to get working.
If there is a better way, please let me know.

https://trello.com/c/p3BVSpj6/740-2-change-banner-when-all-drafts-complete