Set G12 recovery closing date
We now have a confirmed closing date for the G12 recovery phase.
Set this so we show the correct date in the reminder banner.

When we set this previously, the countdown in the banner caused the
visual regression tests to fail the first time they were run everyday.
However, the banner logic was changed in #1331 and our test supplier
in Preview and Staging has one G12 draft service marked as complete.
This means they'll see the static banner content `You have marked 1
service as complete...` and won't get the countdown.

Since #1331 is already released, this commit should not cause any
further visual regression test failures.

https://trello.com/c/LyhdZKkZ/724-1-agree-g12-recovery-dates-and-times