EXPERIMENTAL: lazily load manifests for old frameworks
Currently, it takes ~20 seconds to import all the content. This is a problem because it means that running the very smallest unit test includes ~20 seconds of content importing. This makes running the unit tests slower, which adds up across all developers.

Experiment with lazily importing manifests for old frameworks (prior to DOS5/G12). They will only be imported when needed.

From my testing, this seems to halve the time taken to import all the content, reducing it to less than 10 seconds. This will save developers a lot of time, and also make deploys a little bit faster.

The trade-off is that the first user who needs a manifest for an old framework will have their request take ~1s longer. I think this is OK because users don't access information about old frameworks very often.

This is experimental - I'm creating this for feedback from other developers. If people think it's a good idea, I'll port these changes to content-loader.