Update task list styles
This makes our task list styles more consistent with the Design System coded example.

One variation is that the task list as coded is designed to wrap an `<ol>` around some `<ul>`s, and that's
not the case here, so we can add in some padding overrides so it doesn't look wacky.