Bump govuk-country-and-territory-autocomplete to 0.5.0
This appears to be the last version which is compatible with the functional tests. It appears later versions use different javascript which requires a further interaction after entering a value e.g. click on dropdown value or press return key for the input to be accepted. It is unclear why this was implemented or if it is intentional. I can get this to work using Firefox driver however, I was unable to get it to work using PhantomJS.