Use govukRadios for declaration reuse question
Part of our work to use GOV.UK components instead of our custom ones to improve accessibility.

Switch to use a standard `RadioField`. Previously, we treated it as a checkbox and allowed users to submit without choosing an option. Users are now forced to choose an option, which should reduce the chance of them accidentally not copying a previous declaration and created a lot of unnecessary work.

Also change the page title and header to be the question. This allows us to remove the question as a separate header, which looked a bit odd.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>