Rewrite reuse declaration page
After discussing with @NoraGDS we've decided to change the layout back
to what it was, with a separate page heading and question label, and
the text not associated with the question as a hint. We think that this
is better because the question is directly before the inputs. Discussed
with @domoscargin as well and we agreed that having the question label
bold is good. So we're back where we started, but at least now I know
why we started there :D