Bump govuk-country-and-territory-autocomplete
We want to keep up to date with changes in the data. We previously avoided updating past 0.5.0 in #1355 because it caused
functional tests to fail.

However, it looks like we have a test that works with versions past 0.5.0: alphagov/digitalmarketplace-functional-tests#622

We can use this test in the tests for this form. This doesn't address potential issues like a user typing the full answer
into the autocomplete: alphagov/accessible-autocomplete#472