Replace radio on supplier trading status page
As part of moving to Design System 3, we need to replace our use of `toolkit/forms/selection-buttons.html` with `govukRadios` from the design system. This should make this page more accessible.

As part of this, we're also moving away from our customised DMRadioField to the standard wtforms equivalent.

We've added a helper method in `CompanyTradingStatusForm` to get the items in the right format for `govukRadios`. We're not sure this is the right way to do this - we'll refactor as necessary when we come to replace the other radios in the supplier details journey.

Co-authored-by: Alex Smith <alex.smith@digital.cabinet-office.gov.uk>