Add href to errors
In Design System v2, the first option in a radios component has a `-1` suffix. This is not the case in
Design System v3. We've accounted for this with an `href` property on question objects, but we weren't
using it for the declaration section pages.

This adds that property in. If it doesn't exist, our content loader method will default to using
`input-(input_name)`