Use govuk_options helper to get options for radios
Laurence pointed out that we already have a helper for generating the items for govuk forms. So switch from our custom helper to the existing one.

Required changing the handling of the pre-fill data in edit_supplier_organisation_size to match whgat we already do in edit_supplier_trading_status. This has no impact on what users see, but ensures that everything has the correct type in the tests.