Ask for registration number with govukRadios
As part of moving to Design System 3, we need to replace our use of toolkit/forms/selection-buttons.html with govukRadios from the design system. This should make this page more accessible.

This question is quite complex, so the template is quite lengthy. However, quite a bit of that is just caused by the govukRadios being a bit more verbose and having more braces.