Add support for localstack S3 subdomains
This matches what we use for the real S3 - the old way of doing it in the path is deprecated.

By using localstack S3 subdomains, we now get file upload working correctly. This is because we no longer lose the bucket name when we add the document path, which was the previous behaviour.

This affects supplier-frontend only when run locally in developer mode. It depends on having a recent version of localstack - see https://github.com/alphagov/digitalmarketplace-runner/pull/194. This change means that the functional tests tagged as '@file-upload' now succeed when run locally.