Move S3 endpoint determination into app
In https://github.com/alphagov/digitalmarketplace-utils/pull/580, we added special-case handling for the localstack S3 endpoint into dmutils. This proved to be surprising, because it wasn't obvious why we were setting up the `DM_S3_ENDPOINT_PORT` in the application config.

So move the logic into the app. If we're in an environment where `DM_S3_ENDPOINT_URL` isn't set, then we'll just pass in `None`. That's OK, because boto will use the default endpoint as it already does.