Replace completed service declaration warning banner
We were previously using a `type='warning'` toolkit banner here. The only difference is that it was styled red.

I think this is too easily confused with an Error Summary, and may confuse the user into looking for a validation error on the page.

digitalmarketplace-govuk-frontend also has the "Alert" component, which is styled red, but again, that's used in reaction
to a user's actions - this banner refers to another part of the app.

It's not an ideal fix. Presumably we made the choice to allow users to start adding services before their declaration was complete
as a convenience. If not, it'd be better to make adding serviceables not "startable" until the declaration was complete.