Fix tests for updates table
Previously, this test was broken in two ways:
1. The CSS styles used in the xpaths were no longer in use due to https://github.com/alphagov/digitalmarketplace-supplier-frontend/pull/1286/
2. The test didn't fail when the xpath search found no matching tables.

So I needed to fix this by updating the xpaths to match the current naming. I then updated the tests to fail if it ever fails to find the two tables - for clarifications and communications.

The two affected tests seemed very similar, so I consolidated them into one.