Remove redundant checks and API requests
The `@EnsureApplicationCompanyDetailsHaveBeenConfirmed` annotation already ensures that suppliers have registered interest in a framework and returns a 404 status if not.

So the `if not get_supplier_framework_info(data_api_client, framework_slug):` checks are entirely redundant, since it is imossible for them ever to fail. So we can safely remove them all to remove some unnecessary API requests.

Add a test for this behaviour to the annotation tests. This means that we can safely remove the tests for this behaviour from the individual endpoints.

The history of this is that the annotation was added later (c29a4ee2597b81ce8062b385105479efb5379721). It looks as if they didn't notice that they'd made this check redundant at the time.