Special-case checkbox_tree and pricing input questions
We haven't decided how we're going to deal with checkbox_tree questions,
which we currently only use for selecting service categories. This is an
area which requires some research and design to come up with something
better.

In the meantime, we can special-case this question type so that it doesn't
fail to render.

When we have a solution, we can revert this commit.

Additionally, we have not yet accounted for multiple field pricing inputs.