Fix list of live services for DOS
In 2c265894a847ef947456f5c1e16ab31242134579, I accidentally removed some logic that is needed for the DOS version of this table to render correctly. This is because some DOS lots are limited to one service, so the service has no name.

Fix the logic to match the previous logic, and then add a test for this scenario