Change URL stored for declaration documents
Declaration documents are stored in the documents bucket, the contents of which are already public. So we don't need to go via supplier frontend to access the files. As far as I can tell, this functionality was added as a bug fix. It copied existing code for viewing files in non-public buckets, seemingly without realising that this was unnecessary (https://github.com/alphagov/digitalmarketplace-supplier-frontend/pull/1023).

Instead, use the assets URL to construct the URL we save in the declaration in the database. This will mean that other apps can now show these files to non-supplier users without having to do hacky URL-rewriting.

This also has the added benefit that declaration files uploaded in non-prod environments will now be viewable in buyer-frontend. This is because the current hacky rewriting that this PR makes redundant only works for production.

This won't have an effect on existing files. However, it should take effect for all future frameworks.