Remove 'a person' pricing option for G-Cloud services
In G-Cloud 13 we will only allow unit pricing for lots 1, 2 and 3. There
will be a separate lot 4 that handles cloud support billed on a per
person basis. Lot 4 doesn't exist yet, so we can remove the option to
select per person pricing from the existing lots 1, 2 and 3 and
re-introduce it for lot 4 when that's created.

This only affects the list of pricing options that's shown to the user
when creating a service, not what the API will accept or any existing
services. This page is also only accessible when creating a service,
so there will be no user-facing impact.